chore: remove support contract reference from template (#510)
If a person has made it to GitHub, they almost certainly have an issue with the
library and don't need to circle back to support. By removing this phrase, we
don't confuse customers and make it clear that we're happy to engage here. If a
customer does in fact have a support issue, we can direct them accordingly.