feat: add support for PSC
AlloyDB will soon support PSC-based connections. This commit adds
support for PSC in the Go Connector by exposing an option to Dial. When
the option is provided, the Go Connector retrieves the PSC DNS name from
the connection info response and then attempts to connect to that DNS
name.

Presently, PSC requires some additional setup on the caller's part
(e.g., creation of a Cloud DNS managed zone, with an A record pointing
to the PSC service attachment). Those configuration details will be
provided elsewhere and not here.