fix: remove duplicate refresh for all connections (#526)
When we added a check to validate certificates on all new connections
[1], we also inadverantly caused each connection to trigger a duplicate
and blocking refresh operation with the following lines:

if c.Certificates[0].Leaf == nil {
	return true
}

When the tls.Config is created in refresh.go, it configures a client
certificate chain and adds the CA cert as a Root CA. However, the client
certificate chian is created with tls.X509KeyPair which explicitly
states:

"On successful return, Certificate.Leaf will be nil because the parsed
form of the certificate is not retained." [2]

As a result, the nil check above was always triggered and thereby
causing a duplicate refresh operation. An immediate fix to this
situation is to set the leaf certificate, which is what this commit
does. In fact, we're already parsing the leaf certificate to determine
its expiration and saving the parsed certificate has a performance
benefit:

[The Leaf certificate] may be initialized using x509.ParseCertificate to
reduce per-handshake processing. [3]

[1]: https://github.com/GoogleCloudPlatform/alloydb-go-connector/pull/456
[2]: https://pkg.go.dev/crypto/tls#X509KeyPair
[3]: https://pkg.go.dev/crypto/tls#Certificate.Leaf

Fixes #525.