refactor: simplify internal interface
Rather than construct a tls.Config in the API client, this commit
introduces a ConnectionInfo struct to hold all connection information
including certificates, root CA certs, and expiration. Later those
fields are used to construct a tls.Config type in the same function that
establishes the TLS connection.

This change allows for easier configuration of server name values in the
tls.Config and is in preparation for supporting PSC (which will use the
PSC-based DNS name).