feat: add support for a lazy refresh
When creating a Dialer with the WithLazyRefresh option, the connection
info and ephemeral certificate will be refreshed only when the cache
certificate has expired. No background goroutines run with this option,
making it ideal for use in Cloud Run and other serverless environments
where the CPU may be throttled.

This is a port of
https://github.com/GoogleCloudPlatform/cloud-sql-go-connector/pull/772

Fixes #549