chore: resolve race condition in test assertion (#585)
In some cases, the refresh ahead cache may have already begun a refresh
cycle before the test calls Close on the instance. When this happens,
the returned error is not a context.Canceled error, but instead a
DialError.

This commit updates the TestClose method to check that either error is
returned. The point of the test is to ensure close results in an error
when the next call to ConnectionInfo is made. Which error is less
important.

Fixes #444