refactor: rename internal trace package to tel (#638)
Before introducing OpenTelemetry into this project, it's important to
have a clear distinction between the existing telemetry based on
OpenCensus and the new telemetry which will be based on OpenTelemetry.
To do that, this commit renames the existing package containing metrics
and tracing to be more generic. Later commits will introduce a "tel/v2"
to make the relationship between old and new clear.