chore: add MetricRecorder implementation (#664)
This commit adds the implementation for writing a number of system
metrics to the alloydb.googleapis.com namespace using OpenTelemetry.
Those metrics are:

- alloydb.googleapis.com/client/connector/dial_count
- alloydb.googleapis.com/client/connector/dial_latencies
- alloydb.googleapis.com/client/connector/open_connections
- alloydb.googleapis.com/client/connector/bytes_sent_count
- alloydb.googleapis.com/client/connector/bytes_received_count
- alloydb.googleapis.com/client/connector/refresh_count

The system metrics are public writable, but not public readable yet.

Typically, users of alloydbconn will configure their own OpenTelemetry
exporter, but in this case, the library configures an exporter
internally, such that the system metrics are automatically written
without user configuration.  Presently, it is necessary that the
associated IAM principal that runs this library have the
roles/monitoring.metricWriter role for the writes to succeed.

In terms of Cloud Monitoring's API and the opentelemetry-operations-go
exporter, a metric is attached to a monitored resource. The monitored
resource in this case is alloydb.googleapis.com/InstanceClient and must
include the project, location, cluster, and instance name, in addition
to a unique client identifier. With the exception of the client ID, the
alloydbconn.Dialer does not have this information until the Dial method
is called. As a result, the library cannot configure the Cloud
Monitoring exporter ahead of time. In addition, the
opentelemetry-operations-go library doesn't support attaching labels to
a monitored resource after emitting a metric.

As a result, the code here is designed to create a MetricReporter for
each dialed instance. In practice, this will only matter when one dialer
connects to multiple instances, and even then, the resulting resource
usage will be minor enough to not cause any problems. Nonetheless, it
would be useful to request that opentelemetry-operations-go support
passing in the Cloud Monitoring client to the exporter, to avoid the
need to create multiple clients for the same service and and reduce
resource usage.