feat: allow for disabling built-in metrics (#665)
This commit does two things:

1. It configures the alloydbconn internals to send telemetry to AlloyDB
   reporting on the library's performance (e.g., dial count, refresh
   count, bytes tx and rx)
1. It allows callers to disable this internal telemetry.

Presently, the telemetry is not publicly readable. It is put in place in
this commit as a steel thread for improving the performance,
reliability, and diagnostics for the connector.

Future work will surface these system metrics alongside other AlloyDB
metrics for callers to use as they see fit. Unlike the existing
OpenCensus based implementation which uses custom metrics (paid for by
the caller), this implementation uses OpenTelemetry for system metrics
(paid for by AlloyDB).