fix: shut down the internal exporter only once
The internal metrics use an exporter and a provider. Previously, when
calling Close on the alloydbconn.Dialer, the code would shut down the
provider and the exporter. Meanwhile, the provider also tries to shut
down the exporter, leading to cases where the exporter's shutdown method
was called twice.

This commit updates the Shutdown method on the internal metric reporter
to shutdown only the provider (which shuts down the exporter). In
addition, since internal metric shutdown errors are unactionable, this
commit changes the Close method on the Dialer to log the error when
debug logging is enabled and otherwise not surface it to callers.

Fixes #776