fix: return error with conflicting credential opts (#705)
This commit ensures conflicting credential options result in a
user-facing error.

This commit moves all the dialer config initialization into a single
function in options.go. This will help the migration to supporting
auth.Credentials and makes the relationship between all the options
clear. Previously, there was a mix of option functions setting various
fields, possibly returning errors, and the dialer initializer doing
additional configuration work. This made it hard to understand how the
different credential options related to one another and hid the bug of
allowing callers to set conflicting options.

In addition, this commit adds end to end tests to verify the
credential-related options work as intended. This is a belated port of
https://github.com/GoogleCloudPlatform/cloud-sql-go-connector/pull/460.

Related to #646