feat: add support for per-dial IAM AuthN config
This commit adds a new option WithDialIAMAuthN that allows callers to
enable or disable IAM AuthN on a per-dial basis, regardless of how the
underlying Dialer is configured.

Most clients will configure IAM AuthN on a dialer basis and not need
this option. However, in some advanced uses, it is necessary to enable
or disable IAM AuthN without recreating a dialer. This new option
supports that use case.